fix: post-review security and correctness fixes
- Security: access tokens now validated against token_version from DB/cache
  on every request — logout immediately revokes live access tokens (not just
  refresh tokens). CurrentUser gains token_version field.
- Security: rate limiter now uses X-Real-IP (set by nginx from \$remote_addr,
  cannot be spoofed) instead of X-Forwarded-For rightmost entry
- Security: untrack FrontEnd/.env.production from git (git rm --cached)
- Bug: _rate_limit_locks dict now cleaned up during IP eviction (memory leak)
- Bug: Mapped[float] corrected to Mapped[Decimal] on all Numeric columns
  (stepup.price, order_items.unit_price/total_price); unused Float imports removed
- Bug: idempotency key uses sha256 instead of md5 (scanner-clean)
- UX: Cart.tsx shows warning banner for unavailable items (backend already
  computed unavailable_items; frontend now reads and displays them)
- Types: CartData gains unavailable_items field; UnavailableCartItem interface added
- Dev: COOKIE_SECURE=False restored in local .env (True belongs only in prod)

+++ BackEnd/app/models/order.py
@@ -1,4 +1,5 @@
-from sqlalchemy import String, Integer, Float, Boolean, DateTime, func, ForeignKey, Index, Enum, text, Numeric
+from decimal import Decimal
+from sqlalchemy import String, Integer, Boolean, DateTime, func, ForeignKey, Index, Enum, text, Numeric
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.db.database import Base
 from datetime import datetime
@@ -105,8 +106,8 @@ class OrderItem(Base):
         index=True
     )
     quantity: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
-    unit_price: Mapped[float] = mapped_column(Numeric(precision=12, scale=2), nullable=False)
-    total_price: Mapped[float] = mapped_column(Numeric(precision=12, scale=2), nullable=False)
+    unit_price: Mapped[Decimal] = mapped_column(Numeric(precision=12, scale=2), nullable=False)
+    total_price: Mapped[Decimal] = mapped_column(Numeric(precision=12, scale=2), nullable=False)
     notes: Mapped[str] = mapped_column(String(255), nullable=True)
     created_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True),
